requirements.txt
fastapi==0.95.1
uvicorn
requests==2.29.0
python-dotenv
